Migrate fetch implementation test
